% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbLog}
\alias{dbLog}
\title{metScanR_DB Update Log}
\format{(data frame) Version number and publication dates.}
\usage{
dbLog
}
\description{
A simple log file for the version and date of database revisions.
More information on specific updates to the database can be found at
\url{https://jaroberti.github.io/metScanR/}.
}
