% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metScanR_DB}
\alias{metScanR_DB}
\title{Worldwide, Environmental Monitoring Station metadata}
\format{(list) Metadata from ~107,000 environmental monitoring
stations, worldwide.  The metadata for each site include: \cr
\cr
\emph{namez} (character) Name of environmental monitoring site as defined by governing
network\cr
\cr
\emph{identifiers} (data.frame) Station identifers of environmental monitoring site as
defined by associated networks\cr
\cr
\emph{platform} (character) Station platform (type). A single station may be a specific
platform but it might be associated with many networks / identifers.\cr
\cr
\emph{elements} (data.frame) Elements (environmental phenomena) measured, along with
associated start and end dates\cr
\cr
\emph{location} (data.frame) Spatial location information (latitude, longitude, country,
elevation) of environmental monitoring site\cr}
\usage{
metScanR_DB
}
\description{
Metadata from ~107,000 environmental monitoring stations
among 219 countries/territories and 18 environmental networks.
Metadata are gathered from a growing number of sources and the database is
continually updated to reflect the increase of information.\cr
}
\keyword{atmospheric}
\keyword{data,}
\keyword{datasets,}
\keyword{environmental}
\keyword{metadata,}
\keyword{meteorology,}
\keyword{weather}
