% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metScanR_terms}
\alias{metScanR_terms}
\title{Environmental metadata terms used within metScanR_DB}
\format{(data frame) Environmental terms and data product
identifiers from environmental monitoring networks, worldwide}
\usage{
metScanR_terms
}
\description{
Environmental metadata terms (e.g., air temperature, wind speed,
etc.) and data product identifiers used by the environmental monitoring
networks within the metScanR_DB.  Terms are gathered from a
growing number of sources and the terms database is continually updated to
reflect the increase of information.
}
\keyword{atmospheric}
\keyword{data,}
\keyword{datasets,}
\keyword{environmental}
\keyword{metadata,}
\keyword{meteorology,}
\keyword{n-grams}
\keyword{weather,}
