% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metScanR.r
\docType{data}
\name{NEON_masterTraceability}
\alias{NEON_masterTraceability}
\title{NEON Sites:}
\format{A data frame with 80 rows and 17 variables}
\usage{
data(NEON_masterTraceability)
}
\description{
A dataset containing metadata of sites within the National Ecological Observatory Network (NEON):
}
\keyword{datasets}

