% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metScanR.r
\docType{data}
\name{NOAA_NCEI_masterTraceability}
\alias{NOAA_NCEI_masterTraceability}
\title{NOAA sites:}
\format{A data frame with 11151 rows and 17 variables}
\usage{
data(NOAA_NCEI_masterTraceability)
}
\description{
A dataset containing metadata for networks and sites managed by the National Oceanic and Atmospheric Administration (NOAA):
}
\keyword{datasets}

