% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metScanR.r
\docType{data}
\name{NRCS_masterTraceability}
\alias{NRCS_masterTraceability}
\title{NRCS sites:}
\format{A data frame with 2217 rows and 17 variables}
\usage{
data(NRCS_masterTraceability)
}
\description{
A dataset containing metadata for networks and sites managed by the National Resources Conservation Service (NRCS):
}
\keyword{datasets}

