% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSiteFinder.R
\name{mapSiteFinder}
\alias{mapSiteFinder}
\title{mapSiteFinder}
\usage{
mapSiteFinder(x)
}
\arguments{
\item{x}{A list output from \code{siteFinder()}}
}
\value{
\code{} A map of resulting environmental monitoring sites found using siteFinder()
}
\description{
A plotting tool that maps the results of \code{siteFinder()}.  This function requires internet connection!
}
\seealso{
Currently none
}
\author{
Josh Roberti \email{jaroberti87@gmail.com} \cr
Lee Stanish \cr
Cody Flagg \cr
Sam Weintraub \cr
Derek Smith
}
\keyword{atmopsheric}
\keyword{atmosphere,}
\keyword{data}
\keyword{data,}
\keyword{environment,}
\keyword{environmental}

