% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDatabase.R
\name{updateDatabase}
\alias{updateDatabase}
\title{Update the metScanR database to the latest version.}
\usage{
updateDatabase()
}
\description{
Updates the metScanR database to the latest version hosted on GitHub at: \url{https://github.com/jaroberti/metScanR}.
When installed, the metScanR package contains only a small (subset) database comprising ~5300 environmental monitoring stations.
This function will update the local version of the database to the most up-to-date version.
}
\examples{
\dontrun{
updateDatabase()
}
}
\seealso{
\link[metScanR]{metScanR_DB}
}
\author{
Robert Lee \email{rhlee@colorado.edu} \cr
}
\keyword{Earth,}
\keyword{air}
\keyword{air,}
\keyword{atmopsheric}
\keyword{atmosphere,}
\keyword{canopy,}
\keyword{chemical,}
\keyword{climate,}
\keyword{cloud,}
\keyword{conductivity,}
\keyword{country,}
\keyword{data,}
\keyword{environment,}
\keyword{environmental}
\keyword{flux,}
\keyword{groundwater,}
\keyword{in-situ,}
\keyword{lake,}
\keyword{metadata}
\keyword{phenology,}
\keyword{pollution,}
\keyword{precipitation,}
\keyword{radiation,}
\keyword{reservoir,}
\keyword{river,}
\keyword{salinity,}
\keyword{snow,}
\keyword{soil,}
\keyword{soils,}
\keyword{stream,}
\keyword{temperature,}
\keyword{vapor,}
\keyword{water,}
\keyword{weather,}
\keyword{wind,}
\keyword{world,}
